backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# Security
pyjwt>=2.10.0
argon2-cffi>=23.1.0
cryptography>=44.0.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Excel handling
pandas>=2.2.3
openpyxl>=3.1.5

# CORS
python-dotenv>=1.0.1
